move extract methods to utils
